move employees component into feature module
