+++ src/app/app-routing.module.ts
@@ -1,8 +1,5 @@
 import {NgModule} from '@angular/core';
 import {Routes, RouterModule} from '@angular/router';
-import {EmployeesComponent} from './employees/employees.component';
-import {AddEmployeeComponent} from './employees/add-employee/add-employee.component';
-import {ManageEmployeeComponent} from './employees/manage-employee/manage-employee.component';
 import {SubmitTimeReportComponent} from './time-reports/submit-time-report/submit-time-report.component';
 import {LoginComponent} from './core/auth/login.component';
 import {LoggedOutComponent} from './core/auth/logged-out.component';
@@ -10,19 +7,17 @@ import {AuthGuard} from './core/auth/auth.guard';
 import {HomeComponent} from './home/home.component';
 
 const routes: Routes = [
-  // Default page...
   {path: '', component: HomeComponent},
+  {path: 'submit-time-report', component: SubmitTimeReportComponent},
+  {path: 'submit-time-report/:id', component: SubmitTimeReportComponent},
+  {path: 'login', component: LoginComponent},
+  {path: 'logged-out', component: LoggedOutComponent},
 
-  // Main App
-  {path: 'employees', component: EmployeesComponent, canActivate: [AuthGuard]},
+  // Feature Modules
   {
-    path: 'employees/add',
-    component: AddEmployeeComponent,
-    canActivate: [AuthGuard]
-  },
-  {
-    path: 'employees/:employeeNumber/manage',
-    component: ManageEmployeeComponent,
+    path: 'employees',
+    loadChildren: () =>
+      import('./employees/employees.module').then((m) => m.EmployeesModule),
     canActivate: [AuthGuard]
   },
   {
@@ -32,18 +27,7 @@ const routes: Routes = [
         (m) => m.TimeReportsModule
       ),
     canActivate: [AuthGuard]
-  },
-  {
-    path: 'employees/:employeeNumber/time-sheets',
-    redirectTo: 'employees/:employeeNumber',
-    canActivate: [AuthGuard]
-  },
-
-  // Available to Anonymous
-  {path: 'submit-time-report', component: SubmitTimeReportComponent},
-  {path: 'submit-time-report/:id', component: SubmitTimeReportComponent},
-  {path: 'login', component: LoginComponent},
-  {path: 'logged-out', component: LoggedOutComponent}
+  }
 ];
 
 @NgModule({
