+++ src/app/employees/store/employee.selector.ts
@@ -1,3 +1,3 @@
-import {AppState} from '../app.state';
+import {AppState} from '../../core/app.state';
 
 export const selectEmployees = (state: AppState) => state.employees;
